src/main/webapp/app/entities/task-5/delete/task-5-delete-dialog.component.spec.ts
jest.mock('@ng-bootstrap/ng-bootstrap');

import { ComponentFixture, TestBed, inject, fakeAsync, tick } from '@angular/core/testing';
import { HttpResponse } from '@angular/common/http';
import { HttpClientTestingModule } from '@angular/common/http/testing';
import { of } from 'rxjs';
import { NgbActiveModal } from '@ng-bootstrap/ng-bootstrap';

import { Task5Service } from '../service/task-5.service';

import { Task5DeleteDialogComponent } from './task-5-delete-dialog.component';

describe('Task5 Management Delete Component', () => {
  let comp: Task5DeleteDialogComponent;
  let fixture: ComponentFixture<Task5DeleteDialogComponent>;
  let service: Task5Service;
  let mockActiveModal: NgbActiveModal;

  beforeEach(() => {
    TestBed.configureTestingModule({
      imports: [HttpClientTestingModule, Task5DeleteDialogComponent],
      providers: [NgbActiveModal],
    })
      .overrideTemplate(Task5DeleteDialogComponent, '')
      .compileComponents();
    fixture = TestBed.createComponent(Task5DeleteDialogComponent);
    comp = fixture.componentInstance;
    service = TestBed.inject(Task5Service);
    mockActiveModal = TestBed.inject(NgbActiveModal);
  });

  describe('confirmDelete', () => {
    it('Should call delete service on confirmDelete', inject(
      [],
      fakeAsync(() => {
        // GIVEN
        jest.spyOn(service, 'delete').mockReturnValue(of(new HttpResponse({ body: {} })));

        // WHEN
        comp.confirmDelete(123);
        tick();

        // THEN
        expect(service.delete).toHaveBeenCalledWith(123);
        expect(mockActiveModal.close).toHaveBeenCalledWith('deleted');
      }),
    ));

    it('Should not call delete service on clear', () => {
      // GIVEN
      jest.spyOn(service, 'delete');

      // WHEN
      comp.cancel();

      // THEN
      expect(service.delete).not.toHaveBeenCalled();
      expect(mockActiveModal.close).not.toHaveBeenCalled();
      expect(mockActiveModal.dismiss).toHaveBeenCalled();
    });
  });
});
